requirements.txt
google-cloud-firestore
python-telegram-bot
feedparser
streamlit
folium
requests
pandas
functions-framework
